+++ arangod/Aql/CMakeLists.txt
@@ -71,9 +71,6 @@ add_library(arango_aql STATIC
   NonConstExpressionContainer.cpp
   Optimizer.cpp
   OptimizerRules.cpp
-  OptimizerRulesFeature.cpp
-  OptimizerRulesOptionsProvider.cpp
-  OptimizerUtils.cpp
   OutputAqlItemRow.cpp
   Parser.cpp
   Projections.cpp
