+++ arangod/Aql/ExecutionNode/MultipleRemoteModificationNode.cpp
@@ -33,7 +33,6 @@
 #include "Aql/Executor/ModificationExecutorHelpers.h"
 #include "Aql/Executor/ModificationExecutorInfos.h"
 #include "Aql/Executor/MultipleRemoteModificationExecutor.h"
-#include "Aql/OptimizerRulesFeature.h"
 #include "Aql/types.h"
 #include "Basics/Exceptions.h"
 #include "Basics/StaticStrings.h"
