+++ arangod/Aql/Optimizer/CMakeLists.txt
@@ -1,1 +1,6 @@
 add_subdirectory(Rule)
+
+target_sources(arangoserver PRIVATE
+  Feature.cpp
+  OptionsProvider.cpp
+  Utils.cpp)
