+++ arangod/Aql/Optimizer/Feature.h
@@ -29,7 +29,7 @@
 
 #include "ApplicationFeatures/ApplicationFeature.h"
 #include "Aql/OptimizerRule.h"
-#include "Aql/OptimizerRulesOptions.h"
+#include "Aql/Optimizer/Options.h"
 
 namespace arangodb {
 namespace aql {
