+++ arangod/Aql/Optimizer/Utils.cpp
