+++ arangod/Aql/Optimizer/Utils.h
