+++ arangod/Aql/OptimizerRules.h
@@ -30,7 +30,6 @@
 #include "Aql/ExecutionNode/RemoteNode.h"
 #include "Aql/ExecutionNode/ScatterNode.h"
 #include "Aql/ExecutionPlan.h"
-#include "Aql/OptimizerRulesFeature.h"
 #include "Containers/SmallUnorderedMap.h"
 #include "VocBase/vocbase.h"
