+++ arangod/ClusterEngine/ClusterEngine.cpp
@@ -24,7 +24,6 @@
 #include "ClusterEngine.h"
 
 #include "ApplicationFeatures/ApplicationServer.h"
-#include "Aql/OptimizerRulesFeature.h"
 #include "Basics/Exceptions.h"
 #include "Basics/Result.h"
 #include "Basics/StaticStrings.h"
