+++ arangod/RocksDBEngine/RocksDBOptimizerRules.cpp
@@ -39,8 +39,8 @@
 #include "Aql/IndexHint.h"
 #include "Aql/Optimizer.h"
 #include "Aql/OptimizerRule.h"
-#include "Aql/OptimizerRulesFeature.h"
-#include "Aql/OptimizerUtils.h"
+#include "Aql/Optimizer/Feature.h"
+#include "Aql/Optimizer/Utils.h"
 #include "Aql/Query.h"
 #include "Basics/StaticStrings.h"
 #include "Containers/FlatHashSet.h"
