+++ tests/IResearch/IResearchAnalyzerFeatureTest.cpp
@@ -42,7 +42,7 @@
 #include "Aql/AqlFunctionFeature.h"
 #include "Aql/AstNode.h"
 #include "Aql/Function.h"
-#include "Aql/OptimizerRulesFeature.h"
+#include "Aql/Optimizer/Feature.h"
 #include "Aql/QueryRegistry.h"
 #include "Auth/UserManagerMock.h"
 #include "Basics/files.h"
