+++ tests/IResearch/IResearchViewSortedTest.cpp
@@ -42,7 +42,6 @@
 #include "Aql/ExecutionNode/IResearchViewNode.h"
 #include "Aql/ExecutionPlan.h"
 #include "Aql/ExpressionContext.h"
-#include "Aql/OptimizerRulesFeature.h"
 #include "Aql/Query.h"
 #include "Basics/VelocyPackHelper.h"
 #include "GeneralServer/AuthenticationFeature.h"
@@ -55,6 +54,7 @@
 #include "IResearch/VelocyPackHelper.h"
 #include "Logger/LogTopic.h"
 #include "Logger/Logger.h"
+#include "Aql/Optimizer/Feature.h"
 #include "RestServer/AqlFeature.h"
 #include "RestServer/DatabaseFeature.h"
 #include "RestServer/DatabasePathFeature.h"
